refactor: split out `getFileEntityFromUrl` method
